refactor(mbk/inquiries): split PublicInquiryForm.tsx (934 LOC) into per-step components
Extract form state machine, listing panel, contact section, background
section, field wrapper, and success step into
apps/mybookkeeper/frontend/src/app/features/public-inquiry/.
Orchestrator at pages/PublicInquiryForm.tsx drops to 104 LOC.
All 14 PublicInquiryForm tests pass unchanged.
